Rename PoseidonHasherChip to PoseidonHasher (#113)
