+++ halo2-base/src/poseidon/hasher/spec.rs
@@ -1,4 +1,4 @@
-use crate::{poseidon::mds::*, utils::ScalarField};
+use crate::{poseidon::hasher::mds::*, utils::ScalarField};
 
 use poseidon_rs::poseidon::primitives::Spec as PoseidonSpec; // trait
 use std::marker::PhantomData;
